透明度参数改为整数, 小数点异常情况太多

+++ FloatDock/FloatDockVC/entity/AppInfoEntity.h
@@ -25,7 +25,7 @@ static NSString * AppInfoUrl_Finder = @"file:///System/Library/CoreServices/Find
 @protocol AppInfoArrayEntity;
 @interface AppInfoArrayEntity : JSONModel
 
-@property (nonatomic        ) CGFloat   windowAlpha;
+@property (nonatomic        ) NSInteger windowAlpha;
 @property (nonatomic        ) NSInteger appIconWidth;
 
 @property (nonatomic, strong) NSMutableArray<AppInfoEntity> * windowArray;
